parte 3/requirements.txt
pydantic>=2.5,<3
aio-pika>=9.4
grpcio>=1.60

